feature: choose efficient hash functions when choosing the winner
MD5 and SHA1 are good for our case. Both were designed to check the consistency
of large data and they are very fast. These algorithms were not designed to protect
short password strings. They are far too weak for that. But our case is different we
don't store passwords or other sensitive information. Also, the lowest 16 bits are
different even if we change one byte of the whole lottery, while at the others the
lowest 16 bits remain the same if we have small changes in the data.
Between MD5 and SH1, I have chosen MD5 because is faster than SHA1.