feature: implement custum AnteHandler
DeductFeeDecorator is a custom AnteHandler that checks whether fees are provided.
If the fees are not provided for transaction EnterLottery then an error is returned.
Also, it sends fees to the Lottery module. Latter if the winner doesn't set the highest
bet the fees are transferred from the Lottery module to the FeeCollector module.